+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `flashMessage.onError.message` option in `superForm` renamed to `flashMessage.onError.flashMessage`.
 - `constraints` are now optional in the `SuperValidated` type, and are returned only when loading data, not posting. This is only relevant when you modify constraints before calling `superForm`.
 - Removed the `defaultValidators` option, `'clear'` can now be set directly on `validators` instead.
-- Removed the `emptyIfZero` setting from `numberProxy` and `intProxy`. Use the `empty` option instead, which now has the added `'zero'` setting.
+- Removed the `emptyIfZero` setting from `numberProxy` and `intProxy`.
 
 ### Added
 
@@ -39,3 +39,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `zeroIfEmpty` option added to `intProxy` and `numberProxy`, to handle empty text inputs better.
 - `taintedMessage` can now be an (async) function returning `true` if navigation should happen despite the form being tainted.
 - `onChange` event, that returns a list of modified fields whenever `$form` is updated.
+- Added `'zero'` to the empty option of `numberProxy` and `intProxy`.